create an AuthInfoParser interface implemented by JWTAuthInfoParser
Also reworks ServiceOptions to make it clearer how it should be used by
providing a ServerOptions() method for configuring Twirp servers:

```
	server := elephantine.NewAPIServer(a.logger, a.p.Addr, a.p.ProfileAddr)

	opts, err := elephantine.NewDefaultServiceOptions(
		a.logger, a.p.AuthInfoParser, a.p.Registerer,
		elephantine.ServiceAuthRequired,
	)
	if err != nil {
		return fmt.Errorf("set up service options: %w", err)
	}

	checkServer := spell.NewCheckServer(a, opts.ServerOptions())
	dictServer := spell.NewDictionariesServer(a, opts.ServerOptions())

	server.RegisterAPI(checkServer, opts)
	server.RegisterAPI(dictServer, opts)
```

The logging hook now assumes that AuthInfo is used for authentication context,
this removes the need for the `scopesFunc` argument to AddLogginghooks(). This
together with assuming that authentication information always is passed through
the "Authorization" header lets us keep the AuthInfoParser to a single method.