have the elephantine errgroup capture panics
This was handled in the underlying x/sync/errgroup.Group, but lead to issues. The
elephantine ErrGroup isn't meant to be used as a low level primitive in the same
way as errgroup.Group. ErrGroup is meant to be used "top level" subsystems in a
service, so treating a panic as a normal error should be fine. Documenting the
behavior and intended use.